+++ src/plugins/intel_npu/src/compiler_adapter/CMakeLists.txt
