+++ src/plugins/intel_npu/src/compiler_adapter/include/custom_stream_buffer.hpp
