+++ src/plugins/intel_npu/src/compiler_adapter/include/graph_transformations.hpp
