+++ src/plugins/intel_npu/src/compiler_adapter/include/plugin_compiler_adapter.hpp
