+++ src/plugins/intel_npu/src/compiler_adapter/include/zero_compiler_in_driver.hpp
@@ -10,7 +10,9 @@
 
 #include "intel_npu/utils/logger/logger.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
+#include "izero_compiler_in_driver.hpp"
 #include "zero_executor.hpp"
+
 
 namespace intel_npu {
 
@@ -39,23 +41,6 @@ using SerializedIR = std::pair<size_t, std::shared_ptr<uint8_t>>;
 #define UseCopyForNativeBinary(T)                                                                       \
     (T == ZE_GRAPH_EXT_VERSION_1_2 || T == ZE_GRAPH_EXT_VERSION_1_3 || T == ZE_GRAPH_EXT_VERSION_1_4 || \
      T == ZE_GRAPH_EXT_VERSION_1_5 || T == ZE_GRAPH_EXT_VERSION_1_6)
-
-class ILevelZeroCompilerInDriver {
-public:
-    virtual ov::SupportedOpsMap query(const std::shared_ptr<const ov::Model>& model, const Config& config) const = 0;
-    virtual ze_graph_handle_t compile(const std::shared_ptr<const ov::Model>& model, const Config& config) const = 0;
-    virtual ze_graph_handle_t parse(const std::vector<uint8_t>& network, const Config& config) const = 0;
-
-    virtual NetworkMetadata getNetworkMeta(ze_graph_handle_t graphHandle) const = 0;
-
-    virtual _ze_result_t release(ze_graph_handle_t graphHandle) = 0;
-
-    virtual CompiledNetwork getCompiledNetwork(ze_graph_handle_t graphHandle) = 0;
-
-    virtual void setArgumentValue(ze_graph_handle_t graphHandle, uint32_t argi_, const void* argv) const = 0;
-
-    virtual void graphInitialie(ze_graph_handle_t graphHandle, const Config& config) const = 0;
-};
 
 /**
  * Adapter to use CiD through ZeroAPI
