+++ src/plugins/intel_npu/src/compiler_adapter/src/graph_transformations.cpp
