+++ src/plugins/intel_npu/src/compiler_adapter/src/precomp.hpp
